Gather performance data for EP STREAM fragmented with MAX_LOGICAL
Similar to the HPL and EP additions, made a separate directory for perf runs
with MAX_LOGICAL over a symlink of the original tests, which will be skipped
for runs other than performance trials.

Add this new data to the EP STREAM fragmented graph, and the graph to the
MAX_LOGICAL suite.