Gather performance data for the release spectral norm version with MAX_LOGICAL
Similar to the earlier commits, adds a performance directory populated by
symlinks to the files associated with the original version (with the exception
of the .perfexecopts file, which was symlinked from a different spectral norm
version because the release one specified a .dat file name and I was feeling
lazy).  This directory uses MAX_LOGICAL and only does performance runs.  Add the
new data to the spectral norm graph and add the graph to the MAX_LOGICAL suite.