Change skipif files (again) for new performance suite.
I had been a little uncertain about using the presence of --fast and --static
to determine if a performance test was occurring, but hadn't seen an environment
variable that would cover it when I looked in start_test for the handling of
-performance.  That's because the environment variable is assigned when the
variable created by the flag (performance) is checked later on.  We really need
to rewrite start_test.  Anyways, switch the skipifs back to non-executable and
check against this environment variable.

Thanks to Elliot for pointing this out.