Improvements resulting from review.
Move the node and sublocale ID type definitions and the special
sublocale ID constants into common code in module ChapelLocale.  Add a
definition of chpl_getSubloc() to the flat locale model. It contains a
compilerError() call, to catch inadvertent uses of this proc where it
cannot be useful (since there are no sublocales).