Add binary operators to BitVec implementation.
Also, use size_t in place of int, for closer compatibility with std::bitset and the boost
dynamic_bitset.

Added a copy-constructor and the copy operation to support + (set union) and - (set difference).

Updated bb.cpp, copyPropagation.cpp and loopInvariantCodeMotion.cpp, clients of BitVec,
because now the size it returns and the ndata field are unsigned (size_t).