Runs hello.chpl.
Added filters in extractSymbols() to ignore symbols that are not VarSymbols or
ArgSymbols, extern symbols and symbols not of a record type.  Filtering out symbols of
class type also filters out references.  Maybe this is too clever, but at least fits well
with the style of programming currently in place.  If the representation of reference
types changes in the future, this is one place where the code will have to be amended.  At
least that case is covered by a helpful comment.

Moved the initialization of the alias table up into extractSymbols() because the
expectation that each symbol is initialized before being copied (or copied into) through a
move is currently false.

Split off the isConstructor() test as a predicate.

Added a test in processMove(), so we can flag when an uninitialized symbol is copied.
This would probably result in a large number of false positives before flow analysis is
carried out, because each block starts with an empty gen set.

We now filter by the members of the symbolIndex, so we ignore symbols that are not local
(including module-level and extern symbols) and not variables or arguments.

Had to add a test to skip degenerate basic blocks.  Why do we still have degenerate basic
blocks?

We skip functions that are prototypes (including extern functions) altogether.

Re-enabled insertGlobalAutoDestroyCalls() and activated the new insertAutoDestroyCalls()
mini-pass.