Now actually adding autoDestroy calls.
Added a backward flow pass to propagate disharge-of-ownership constraints from IN sets
back across the gap to OUT sets.  The OUT sets of terminal blocks are set to all zeroes to
force all locals to be autodestroyed on exit.

Added code so that we only add an autoDestroy for one member of an alias clique.