Comment out the clause that prevents insertion of autoCopy calls for reference counted types.
Some tests involving internally ref-counted types were failing due to the reference count
going negative.  This turned out to be due to record-wrapped types (the same set) being
inserted into tuples without their copy-constructor (autoCopy) being called.  The
destructor for a tuple calls autoDestroy on its elements, so these autoCopy calls are
required.

It is probably the case that destructor calls were omitted for record-wrapped elements in
a tuple.  But since I'm coding the insertion of destructor calls from scratch and trying
to keep the implementation as uniform as possible, tuple destructors contain destructors
for all of their record elements, and are called wherever a tuple local needs to be
destroyed.

Adding more autoCopies exposed that we may have RWT objects that are empty (i.e. contain a
nil class object) and autoCopy should be robust to that.

291 failures remaining