Initial fix for standalone parallel iterators with arguments
In cleanupLeaderFollowerIteratorCalls() in lowerIterators.cpp, it looks for
functions named "_toLeader", "_toFollower", and "_toFastFollower" to clean up.
Add "_toStandalone" to that list.

Add a new test standaloneArgs.chpl that defines and uses leader/follower and
standalone parallel iterators that take arguments.

In the interest of getting this to a github branch as quickly as possible, I've
only done minimal testing so far.  Tested against the new test,
standaloneParallelIter.chpl and test/release/examples with default linux64.