Update per reviewer recommendations and other tweaks.
Added length compatibility tests in operators that access an "other" BitVec.

Added a length test in get().

Added const specifiers to methods and reference operands, to advertise where the
object and/or operand is not modified.

Abandoned performant donor/adopter pattern for a more understandable version of the dyadic
operators + and - .