Strengthened the test in isBitwiseCopy(), so it returns false if the RHS is global.
Removed the commented-out clause that prevented autoCopy insertion for ref-counted types
(no change in functionality).

Updated some test output because the copy constructor is called for members of tuples much
more often now.

In users/ferguson/refcnt, the output changed because autodestroy calls have been moved
closer to the end of the routine.  This is good, because it shows that in the new
implementation there are more opportunities for an owned variable to be consumed before it
goes out of scope.