Ensure unpack script fails when any command fails in it.
The patch command in the re2 unpack script fail for systems without patch
installed. But, the unpack script did not stop and it did not report the error
(it prints the message, but continues on and eventually exits with zero code),
so the makefile and everything else assumed that re2 built without issue.

Later on when building the runtime, however, compiling re2-interface.cc fails
because it cannot find FilePiece. FilePiece is defined in file_strings.h and is
added to re2 with the patch.

Update the unpack script to fail if any of its commands fail.