Skip io/ferguson/ctests/ when atomics==locks && tasks!=fifo.
These C unittests exercise particular sections of the IO implementation. Some
rely on atomics. When atomics are implemented using locks, the tasking layer is
responsible for implementing the locks.

Due to the way C tests are run, including headers from third-party/ does not
work very well (and is not provided by default). So, in the case that these
tests require headers from the tasking layer implementation (i.e. when atomics
are implemented as locks, as opposed to intrinsics, and the the tasking layer
is not fifo), simply skip the tests.

Note that this does not mean locks do not work in this configuration. Rather,
this is a limitation of the C testing infrastructure.