Fix BB analysis bug: Empty blocks with no predecessors
Added cleanup traversal to remove empty blocks.  (Also control verification with the
fVerify flag now.)  Revised the verify function to lock this behavior in.

The removal of empty blocks cannot be done until BB analysis is complete, because some of
the threading is delayed.  We have to wait for threading to finish before we can identify
blocks that have no predecessor.

Updated the signature of initInSets in copyPropagation.cpp.