Update test/classes/figueroa/RecordConstructor2.good due to increased autoCopy callse.
This test basically devolves to a test on whether the members of a tuple are passed by
reference or value.

In the current version of the grand autoCopy/autoDestroy rework, a tuple owns and destroys
record members that are passed by value.  The output of this test is consistent with that
definition.

In the previous implementation, MM in tuples was something of a hybrid.  (Record) values
copied when the fields in a tuple were created and destroyed when done, but
(inconsistently) the members of a tuple are not copied when a tuple is assigned or when
one of its members is read out.  This is the root cause of the return-record bug this work
is intended to fix.

However, this solution is not necessarily permanent.  It may be possible to avoid these
copies by handling tuples specially.  In that special implementation, tuple members would
always be treated as references.  But if a member of a tuple is used to determine the type
of a result, members that were originally values (i.e. not references) would have value
and not reference type.

There are some semantical issues related to this choice that may need to be worked through
(or experimented with).

Down to 43 regressions.