Changes to chpl-cache for compatibility with NDEBUG
chpl-cache has a few verification routines that use a lot of asserts and
don't make much sense with NDEBUG set. These routines are only called if
VERIFY is set to 1 (defaults to 0), so I've added a few #ifs to prevent
the verification routines from being compiled at all when they are not
used.