Disabled initializeClass.
Nested calls to initializeClass() result in the creation of temporaries which are then
piped through chpl__autoCopy to initialize fields in the containing record or class.  When
the field type is an extern type, the initializeClass() does nothing.  This can be bad in
cases where an initial value (such as nil) is expected due to the class (or record)
invariants.

It turns out that we don't need initializeClass() anyway, since it is used only in
specializations of the compiler-supplied all-fields constructor.  These specializations
should ensure that all fields are initialized, so code generated by initializeClass()
should be moot.  (Testing indicates that this is the case, but I have not yet extensively
studied the generated code.)

I tried earlier to remove the PRIM_INIT_FIELDS primitive entirely, but apparently it is
still needed to establish the type of _this within the specialized all-fields constructor.

I reverted the change to IO.chpl, since it did not actually resolve the problem.

Down to 30 regressions.