Update sub_test to process .skipif files before any others.
There is one test system that does not support symlinks due to its version of
git. In a happy coincidence, that system also skips one of the two failing
tests.

Aside from the above coincident, it generally seems more efficient to evaluate
the .skipif before doing any other work on the test.