Skip filerator/walk test when subdir-link is not a symlink.
On an older test system, the subdir-link is not checked out as a symbolic link
because of how JGit is used (it forces core.symlinks to false). Add new
walk.skipif that checks if subdir-link is a symbolic link and if not returns
True. This should skip the one remaining error case resulting from broken
symlinks in the test system.