Fix a bug in the insertion of autoDestroy calls
Too many autodestroy calls were being inserted because one temp contributing to an alias
clique might lie along one path while another lies along a distinct path.  Backward flow
analysis then considers them both to be dead upon entry to the join node.  The constraint
that each of these temps be dead on exit from their respective blocks causes autodestroy
calls to be inserted.  The temps themselves are dead as required, but the other members of
the clique get deleted (because they are all aliases) even though their respective bits in
the OUT set may be true.  This mismatch causes us to get the wrong answer, assuming that
one of those other clique members is alive when it is actually dead.

To address this, I moved the population of the alias map out to a separate traversal.
Now, any manipulation of the flow sets affects all members of an alias clique in concert,
whether or not their equivalence has yet been established (through a bitwise copy).

The result is that this mode of premature deletion has been eliminated.  This change
allows the insertion of autoDestroys for internally reference-counted types to be
enabled.  Doing so, only cause one additional regression to appear.

Now at 31 regressions.