yield after every ~100 uncontested locks of a sync var (qthreads)
To prevent starvation due to never switching away from a task that is spinning
while doing readXX() on a sync variable we currently do a qthread_yield every
4096 unlocks of a sync var. However, this doesn't seem to be enough and we
still got occasional timeouts with tests like domains/sungeun/assoc/stress.chpl

This patch attempts to solve the same problem in a different manner. This
yields every ~100 uncontested locks of a sync var instead of every 4096 unlocks
of a sync var. There is precedent for this in some of our other tasking layers.
It seems to have the desired effect of eliminating starvation (at least for
1,000 trials of domains/sungeun/assoc/stress.chpl) but I have to check if there
are any performance implications of this.