Turn down the screws in memory/figueroa.
Resolved the remaining leak in LeakedMemory2 by removing extra autoCopies that were being
inserted in chpl__autoDestroy for internally reference-counted types.  In contrast to most
types, internally reference-counted types are passed by value to the destructor.
Autocopies are not applied to ref arguments, but a special case is required to prevent
their application to args and field references in the context of chpl__autoDestroy and
destructor functions when the type of the object being destroyed is an internally
reference-counted or sync type.

479 regressions (no regressions in memory/figueroa/)

###DETAILS
+++ b/compiler/passes/insertAutoCopyAutoDestroy.cpp
Added a flag, so we also consider autoDestroy functions to be consumers.

Renamed isDestructorArg as isDestructorFormal, to more accurately reflect its behavior.
Added a new predicate isDestructorArg, and used this to prevent application of autoCopy to
arguments being passed to destructors.  We assume that at the top level, the argument to a
chpl__autoDestroy function is always owned.  Some of the remaining regressions may be due
to this assumption being violated.

+++ b/compiler/resolution/functionResolution.cpp
In resolveExpr(), we skip resolution of the type constructor for runtime types
(because they are generated on the fly and do not have type constructors defined for
them).

In buildRuntimeTypeInfo(), added the creation of a default constructor and an autoDestroy
function for runtime types.  These were missing, so after all other leaks were cleared up,
LeakedMemory2 was still leaking memory from a copy of a domain that was never getting
released.

+++ b/modules/internal/ChapelArray.chpl
Increment the reference count in _newArray, _newDomain, _getDomain, _newDistribution and
_getDistribution.  Each of these creates a new reference to the type underlying those
record-wrapped type.  To obey the new ownership model, the ref count has to be incremented
when these new wrappers are created (or rather when the wrapped types become referenced by
said wrappers.