Allow empty lines in suppression files and cleanup filterSuppressions
Previously empty lines in a suppression file would cause problems. This was
because we do a regex match against each non-comment line and so we matched
against the empty string and would remove the first failure found incorrectly.

This updates filterSuppressions to use the ReadFileWithComments helper we
already had to avoid this issue. This lets us get rid of some lines that
checked for comments and tried to strip the "real" lines in a suppression file

I ran into this while adding suppressing for baseline regressions