Add memory management for runtime types.
26 of the remaining 30 regressions were memory leaks due to array elements with "runtime
types" not being cleaned up.  That is because runtime type records are added after
resolution, so they don't automatically get autoCopy and autoDestroy functions defined for
them.  The machinery in insertAutoCopyAutoDestroy.cpp is set up to just skip the insertion
of autocopies and autodestroys if there is no entry in the corresponding type->fn map.

I was having trouble with the runtime types being redundantly named.  My Q&D solution was
to number them serially.  When I encountered the naming problem, I had already implemented
most of the machinery for creating full-up runtime types.  But at that time, it occurred
to me that a better approach might be to move the runtime type definitions (as generic
types) out to module code.  Then, most of what I had to code here can be ripped out, and
hopefully much of the runtime type machinery as well.

###DETAILS

####./compiler/AST/expr.cpp
There must be a normalization that moves the contents of the field name (as a literal)
into the the actual name of the field variable.  But runtime types happen after this
normalization happens.  To get codegen to work correctly, we have to punch through at get
the literal string the stores the field anme.  (Maybe this will work for all field names.)

####./compiler/passes/scopeResolve.cpp
In build_type_constructor(), added a clause so we will add generic args from a runtime
type to its type constructor.  Rather than specifying the type of a field using the
exprType field and then resolving this, a runtime type leaves both the init and exprType
fields null and sets the initial type of an element type field to dtUnknown.

####./compiler/resolution/functionResolution.cpp
Added the functions resolveRuntimeTypeDestructor(Type* type),
buildRuntimeTypeAutoDestroyFn(AggregateType* ct) and
buildRuntimeTypeInitCopyFn(AggregateType* ct).

In resolve_type_constructor() added a clause to handle resolution of the type constructor
for a runtime type specially.  Generic arguments are specifically passed along to the type
constructor; others are not.

In buildRuntimeTypeInfo(FnSymbol* fn), a unique name for each runtime type is constructed
on-the-fly.

Added a clause in removeRandomPrimitive(CallExpr* call), so that the field for field-name
substitution is not performed for the type constructors added for runtime types.  This is
related to the field name normalization mentioned above.

In buildRuntimeTypeInitFn(FnSymbol* fn, Type* runtimeType), fields are now added by name.

The function resolveRuntimeTypeDestructor() is no longer needed.  It is resolved when the
autoDestroy function that calls it is resolved.
The function buildRuntimeTypeAutoDestroyFn is not needed, because, we rely on the standard
record autoDestroy function.
However, we now have to supply a RuntimeTypeInitCopyFn.

####./compiler/resolution/wrappers.cpp
In buildDefaultWrapper(), there was a coding error.  We want to use the field name, not
the formal name when generating PRIM_SET_MEMBER calls.