Clean up non-blocking comm support in the comm layer interface.
The part of the comm layer interface that supports non-blocking
communication isn't sufficient to support all uses, and has some
differences with the rest of the interface.  This cleans that up.
The specific changes are as follows.

Rename to get the "verb" first:
  chpl_comm_nb_handle_is_complete() --> chpl_comm_test_nb_complete()
  chpl_comm_nb_wait_some() --> chpl_comm_wait_nb_some()

Add:
  chpl_comm_try_nb_some()

And in comm-gasnet.c, add assert() invocations to ensure that NULL and
GASNET_INVALID_HANDLE are the same.  The code relies on this implicitly,
but it's not stated anywhere and there are no explicit conversions that
would take care of the translation if they were different.