Simplify non-performance PTRANS pass/fail checking.
These versions of PTRANS report both success or failure and the
discrepancy between their result and the expected result range.  Testing
was passing everywhere we ran it, but it passed differently on different
platforms: matching the expected result exactly everywhere but cygwin32
and merely falling within the expected range on cygwin32.  This caused
false failures on cygwin32 due to the .good not matching.

Here we simplify pass/fail checking for testing by adding a config const
that allows limiting the success/failure output to merely whether or not
the result was within the expected range, not how closely.  This will
allow the simple .good files to match on all platforms.