Run deadBlockElimination() near start of pruneResolvedTree().
Currently in postFold() when the conditional expression in a CondStmt can be reduced to
"true" at compile time, the compiler replaces the entire CondStmt with its "then" block.
But the "else" block remains in the tree.  Since the CondStmt has been removed, the
conditional expression is no longer reachable.  In basic block analysis, it becomes a
non-empty block with no predecessors.

This commit runs deadCodeElimination() in pruneResolvedTree() just after
removeUnusedFunctions().  It removes the dead "else" blocks left in the tree by the
folding of param "if"s.

This change is required to support the new implementation of Automatic Memory Management.
AMM relies on basic block analysis to compute whether a variable is owned or not.  Before
the folding of param "if"s, it will find that variables that are initialized prior to the
"if" are owned at the end of both the "then" and "else" clauses.  After a param "if" is
folded, there is no way for flow to get to the "else" clause, so it becomes unreachable.
In that case variables that are owned on entry to the "then" clause will be shown as
unonwed on entry to the "else" clause -- which is an error.

TODO: Try the alternative of fixing up the tree in CondStmt::foldConstantCondition()
instead.

###DETAILS

####./compiler/include/passes.h
Add deadBlockElimination(), and make it a global symbol.

####./compiler/optimizations/deadCodeElimination.cpp
Factor deadBlockElimination(FnSymbol*) into two supporting routines: findReachableBlocks()
and deleteUnreachableBlocks().
Moved the initialization and reporting of the number of dead blocks removed into
deadBlockElimination().
Added a liveness check on functions deadBlockElimination, so the pass can be run
immediately after resolution and before unused functions are garbage-collected.

####./compiler/resolution/callDestructors.cpp
Added a workaround in insertAutoCopyTemps() to ignore symbols that are never defined.
Dead block elimination can remove the DefExpr statement from the tree while the symbol is
used elsewhere (e.g. in a ForLoop statement).  A cleaner version of this would probably
result in the ForLoop being removed as well, but this cleanup has not been fully
implemented.

####./test/optimizations/hilde/removeUnreachable.good
The output of this test changed, because deadBlockElimination() is now called twice.