Remove flwlen==1 check in range follower
I'm guessing this was an "optimization" to avoid creating the range, but the
second yield in here prevents the optimize-loop-iterator optimization from
firing which kills performance.

I'm pretty confident that removing this check doesn't have any correctness
implications. The calculation to find high (for flwlen==1, low must equal high)
is "high = low + stride * (flwlen - 1)" which will just be "high = low +
stride * 0" or "high = low" So correctness wise this is fine. It means we're
building up a range for flwlen = 1, but that overhead is far far far less than
the overhead of the optimize range iterators optimization not firing.

For test/performance/bharshbarg/range-forall.chpl, this drops the execution
time from ~90 seconds to ~2 seconds on chap04. domain-forall.chpl is around
~1.2 seconds.  A future patch will address the remaining gap (by using a
non-strided range in the follower when we can.)

This code was originally added with b35c16e6f2dd124ed5ec0595fa4ec466d382e9ee