Handle attempts to remove fields of ForLoops more gently.
Hacked in missing ForLoop::replaceChild().

This didn't work anyway, since ForLoop::verify expects the mIndex and mIterator fields to
be non-null.  So I revised the conditional code in deleteUnreachableBlocks() to
pick out and reject the specific cases we're tripping over.

TODO: After this revision, removal of dead ForLoops (at least) is incomplete.  Because we
do not remove contained DefExprs, the body of a deaf ForLoop never collapses down to NULL,
and we don't have a pass to harvest such trivial loops anyway.

TODO: The traversal of expressions in dead block removal should be in postorder, so
hollowed-out constructs (like ForLoops) can be identified and removed in one pass.