Add custom value and label formatters for dygraphs
The value and label formatters change the formatting of the values in y-axis
and in the legend. The custom formatters just adjust the number of digits. This
replaces the logic that used to be in the zoom callback to adjust the number of
digits displayed after decimals.

Doing it in the zoom callback required re-rendering the dygraph which is slow.
This should give a substantial performance boost when a graph is first loaded
since we avoid a re-render that we ended up doing on the initial load for most
graphs.  It should also help a little bit with zooming, but not as much since
most of the time the number of decimals didn't change much on zooming.