Replaces creoleparser+html output for chpldocs with rst output.
Pulls all creole references from the docs pass.  This removes quite a number of
fDocsTextOnly flag uses.  Now, the other option is rst and the switch between
these two options is a map lookup - a global map is created that stores pairs
from the chapel structural type to its plaintext or rst mirror.  For instance,
in plain text a record would start its definition with "Record: NAME" while in
rst its definition would start with ".. record:: NAME", so the map key "record"
would lead to "Record: " in plaintext mode and ".. record:: " in rst mode.
Certain structures have a prefix in rst mode that they do not have in plaintext
mode, such as functions/methods: in this case, for rst mode they store
".. function:: " or ".. method:: ", which is concatenated to the output
generated in both plaintext and rst mode.  i.e.

proc foo() // in the Chapel file

turns into:

.. function:: proc foo() // in rst mode
proc foo() // in plaintext mode

with some amount of indentation appropriate for where it is nested.  Likewise,
fields will be prefixed with ".. attribute:: " in rst mode, etc.  This change
does not affect current chpldoc testing, the behavior should be the same.

Our rst mode will pay attention to whether a defined function is a method on a
class/record or whether it is a module level function.  To that end, some logic
needed to be added to printFunctions so that it would know from whence it came.

Because creole is no longer used in the docs pass, I have removed chpldoc2html
and its supporting .css files.