Change the intent on the formal of a serial iterator function to "blank".
Assignment from a serial iterator to an array was segfaulting because the domain iterator
"these" was returning a nude DefaultRectagularDom.  This was being initCopied in the
getIterator function.  The result was copied into an iterator class field (as a class
object, so no autoCopy there) and then destroyed at the end of the function.  After that,
the reference being carried around in the iterator class object was pointing to freed
memory.  Crud.

I thought things might work better if the initCopy was an autoCopy instead, or simply
wasn't there.  I traced the initCopy back to INTENT_CONST_IN on the iterator argument in
the serial iterator function cobbled together in build.cpp.  All of the other 3 iterator
functions built on the fly in that file take their iterator arguments by blank intent.
This was just a land-mine waiting for me.

Long explanation; one-liner fix; 65 regressions remaining.