Scroll the graph selection pane with the performance graphs
Previously the graph selection pane (suite selection, individual graph
selection, etc) stayed stationary and you had to scroll back up to the top of
the page to select different graphs. Now it scrolls as you scroll the page. In
order to do that the individual graph selection list had to be put inside a new
pane so it can scroll if there's more graphs than can fit on the current page.

This adds logic to do the scrolling and automatically adjust the height of the
graph list on resizes and set the initial position of the graphPane based on
the width of the graphSelectionPane