Switch to use fixed positioning for the graph selection pane
Absolute positioning made scrolling on windows really choppy for some reason.
This changes to use fixed positioning.

Previously we manually moved the selection pane vertically to compensate for
scroll. Now the browser automatically moves it, and instead we have manually
move it out of view when scrolling right.

This has the same affect as before, but isn't choppy on windows