Support 'today' in perf graph URL
You can now use 'today' as a sentinel for startdate/enddate in the URL for
performance graphs. This allows you to save/send links that will always contain
the most recent data regardless of when you saved/sent the link. 'today' will
never get set automatically right now. We have discussed having it set if you
zoom in on today's date, but haven't decided if we actually want to do that
yet. The functionality to do so is included here, it's just commented out for
now.

A few supporting changes/fixes:
 - Add a helper function to get the current date
 - Add simple helper/wrapper functions to set/get the date to/from the URL
 - Add simple helper/wrapper function to check if two date ranges are different
 - Update the cleardates function
    - It wasn't quite right if you refreshed a page while already zoomed in. It
      reset graphs back to the date range when they were created, but that
      could be different than the startdate/enddate in the graphinfo if the
      graph was created with an initial date that was from the URL.
 - Update logic for when we set the date in the url when we zoom
    - only update the url if we actually changed the zoom along the x-axis