Fix valgrind regressions from the standalone parallel iterators commit
In some cases, the defaultInitializer function for a type gets removed as
unused at the end of function resolution, but the type's defaultInitializer
field is left pointing at the removed function.  Later, checkResolved looks
at the defaultInitializer for some return types involving iterators.  Before
removing the function, check if it is the defaultInitializer for any type and
set the defaultInitializer field for that type to NULL if it is.

Tested on the regressed tests with valgrind, and the whole suite with
linux64+fifo and linux64+qthreads.